update inputs styles
